refactor responsive options to be more native
