+++ src/Responsive.Core/src/Configurations/Options/ResponsiveOptions.cs
