+++ src/Responsive.Core/src/Configurations/Options/ViewLocationOptions.cs
