+++ src/Responsive.Core/src/CoreCollectionExtensions.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection.Extensions;
 
 using Wangkanai.Responsive;
-using Wangkanai.Responsive.Builders;
+using Wangkanai.Responsive.Configuration;
 using Wangkanai.Responsive.Core.Internal;
 
 namespace Microsoft.Extensions.DependencyInjection
